plugins/ai-security-assessment/skills/ai-security-assess/scripts/pyproject.toml
[project]
name = "ai-security-assess-scripts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
